requirements.txt
Flask
Flask-Cors
python-dotenv
transformers
torch
openai
sqlite3
